Create TCP client and Server problem
his code is just for illustration; you are unlikely to ever write anything so simple.
This does not handle multiple clients well; each client must wait until the previous client is completely served before it even gets accepted.
As in virtually all socket programs, a server socket just listens, and a different, “plain” socket communicates with the client.
The ServerSocket.accept() call is a BLOCKING CALL.
Socket communication is always with bytes; therefore sockets come with input streams and output streams. But by wrapping the socket’s output stream with a PrintWriter, we can specify strings to write, which Java will automatically convert (decode) to bytes.
Communication through sockets is always buffered. This means nothing is sent or received until the buffers fill up, or you explicitly flush the buffer. The second argument to the PrintWriter, in this case true tells Java to flush automatically after every println.
We defined all sockets in a try-with-resources block so they will automatically close at the end of their block. No explicit close call is required.
After sending the datetime to the client, the try-block ends and the communication socket is closed, so in this case, closing the connection is initiated by the server.
Run the server: